feat:send using the rtp dynamic threading version
