+++ .gitignore
@@ -32,3 +32,4 @@ build/
 ### VS Code ###
 .vscode/
 .worktrees/
+.claude/worktrees
