+++ src/main/java/com/cfsl/easymrcp/tts/NettyTtsRtpProcessor.java
@@ -5,193 +5,224 @@ import com.cfsl.easymrcp.mrcp.TtsCallback;
 import com.cfsl.easymrcp.rtp.AudioCodecUtil;
 import com.cfsl.easymrcp.rtp.NettyAudioRingBuffer;
 import com.cfsl.easymrcp.rtp.NettyRtpSender;
-import com.cfsl.easymrcp.utils.SipUtils;
+import com.cfsl.easymrcp.tts.scheduler.TtsProcessScheduler;
+import com.cfsl.easymrcp.tts.scheduler.TtsRtpScheduler;
+import com.cfsl.easymrcp.utils.SpringUtils;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.ByteBufAllocator;
 import io.netty.channel.Channel;
+import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 
-import java.util.Arrays;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
- * TTS音频处理器
- * 负责处理和发送TTS合成的音频数据
- * 使用NettyCircularAudioBuffer提供高性能的环形缓冲区
+ * 共享调度版 TTS RTP 处理器。
+ * 保留当前主链路的 input/output 双缓冲、processOnce()、interrupt() 和共享调度语义。
  */
 @Slf4j
 public class NettyTtsRtpProcessor {
-    private NettyRtpSender sender;
+    /** 输入缓冲时长，单位秒。 */
+    private static final int TTS_INPUT_BUFFER_SECONDS = 30;
+    /** 输出缓冲时长，单位秒。 */
+    private static final int TTS_OUTPUT_BUFFER_SECONDS = 30;
+    /** TTS 链路统一按 8kHz 采样率组织缓冲。 */
+    private static final int SAMPLE_RATE = EMConstant.VOIP_SAMPLE_RATE;
+
+    @Getter
+    private final NettyAudioRingBuffer inputRingBuffer;
+    @Getter
+    private final NettyAudioRingBuffer outputRingBuffer;
+    @Getter
+    private final NettyRtpSender sender;
+
+    private final AtomicBoolean stop = new AtomicBoolean(false);
+    private TtsProcessScheduler processScheduler;
+    private TtsRtpScheduler rtpScheduler;
+
     @Setter
     private TtsCallback callback;
     @Setter
     private String reSample;
-    private int mediaType;
-
-    // 缓冲区配置：支持大容量TTS音频数据
-    private static final int TTS_INPUT_BUFFER_SECONDS = 30;  // 输入缓冲区
-    private static final int TTS_OUTPUT_BUFFER_SECONDS = 30; // 输出缓冲区
-    private static final int SAMPLE_RATE = EMConstant.VOIP_SAMPLE_RATE; // 8kHz
-
-    // 线程间缓冲队列 - 使用NettyCircularAudioBuffer替代RingBuffer
-    private final NettyAudioRingBuffer inputRingBuffer;
-    private final NettyAudioRingBuffer outputRingBuffer;
-
-    private final AtomicBoolean stop = new AtomicBoolean(false);
-
-    // 为解决24khz采样率重采样到8kHz采样率时数据长度不为6的倍数时导致的偶发的噪声问题，统一读取6 * n字节
-    private int receiveTakeBytes = 6 * 500;
-
-    /**
-     * 需要跳过的末尾的字节数(PCM格式)
-     * 例如使用：PCM 8kHz 16bit
-     * 那么200ms = 8000 * 2 * 0.2 = 3200字节
-     */
     @Setter
+    @Getter
     private int skipBytesInTheEndPacket = 0;
+    @Getter
+    private final int mediaType;
+    @Getter
+    private final int frameBytes;
+    @Getter
+    private final int sendIntervalMs;
+    @Getter
+    private final boolean encodeRequired;
 
     /**
-     * 构造函数 - 使用RtpManager
-     *
-     * @param remoteIp   远程IP地址
-     * @param remotePort 远程端口
-     * @param mediaType sdp使用的编码
+     * 为解决 24kHz -> 8kHz 时尾部不对齐带来的噪音问题，沿用旧版固定按 6*n 读取的策略。
      */
-    public NettyTtsRtpProcessor(String remoteIp, int remotePort, int mediaType) {
+    private int receiveTakeBytes = 6 * 1000;
+
+    /** 处理调度任务 ID，接入共享处理调度器后用于取消任务。 */
+    private String processTaskId;
+    /** 发送调度任务 ID，接入共享发送调度器后用于取消任务。 */
+    private String schedulerTaskId;
+
+    public NettyTtsRtpProcessor(String remoteIp, int remotePort, int mediaType, int frameBytes, int sendIntervalMs) throws Exception {
         this.mediaType = mediaType;
-        // 初始化高性能环形缓冲区，TTS模式支持自动扩容
-        ByteBufAllocator allocator = ByteBufAllocator.DEFAULT;
-        this.inputRingBuffer = new NettyAudioRingBuffer(allocator, SAMPLE_RATE, TTS_INPUT_BUFFER_SECONDS, true);  // TTS模式
-        this.outputRingBuffer = new NettyAudioRingBuffer(allocator, SAMPLE_RATE, TTS_OUTPUT_BUFFER_SECONDS, true); // TTS模式
-
-        log.debug("TTS缓冲区初始化完成");
-
+        this.frameBytes = frameBytes;
+        this.sendIntervalMs = sendIntervalMs;
+        this.encodeRequired = mediaType == AudioCodecUtil.PT_PCMA || mediaType == AudioCodecUtil.PT_PCMU;
         try {
-            // 创建RTP发送器
-            this.sender = new NettyRtpSender(remoteIp, remotePort);
-            // 设置Payload Type
-            this.sender.setPayloadType(mediaType);
+            this.processScheduler = SpringUtils.getBean(TtsProcessScheduler.class);
         } catch (Exception e) {
-            log.error("初始化NettyTtsRtpProcessor失败", e);
+            this.processScheduler = null;
         }
+        try {
+            this.rtpScheduler = SpringUtils.getBean(TtsRtpScheduler.class);
+        } catch (Exception e) {
+            this.rtpScheduler = null;
+        }
+        ByteBufAllocator allocator = ByteBufAllocator.DEFAULT;
+        this.inputRingBuffer = new NettyAudioRingBuffer(allocator, SAMPLE_RATE, TTS_INPUT_BUFFER_SECONDS, true);
+        this.outputRingBuffer = new NettyAudioRingBuffer(allocator, SAMPLE_RATE, TTS_OUTPUT_BUFFER_SECONDS, true);
+        this.sender = new NettyRtpSender(remoteIp, remotePort);
+        this.sender.setPayloadType(mediaType);
+        this.sender.configureSession(frameBytes, !encodeRequired);
     }
 
+    /**
+     * 设置 RTP 通道，供外部在媒体链路建立后注入 Netty Channel。
+     */
     public void setRtpChannel(Channel channel) {
         sender.setRtpChannel(channel);
     }
 
     /**
-     * 开始处理音频数据
+     * 上游以 byte[] 形式写入原始 PCM 数据，写入 inputRingBuffer，供 processOnce() 后续处理。
      */
-    public void startProcessing() {
-        new Thread(() -> {
-            while (true) {
-                try {
-                    // 为解决24khz采样率重采样到8kHz采样率时数据长度不为6的倍数时导致的偶发的噪声问题，统一读取6 * n字节
-                    while (inputRingBuffer.getSize() == 0 || inputRingBuffer.getSize() < receiveTakeBytes) {
-                        // 判断是否含有结束标志，有的话取出最后的音频数据并去掉末尾200ms
-                        if (inputRingBuffer.getSize() > 0) {
-                            ByteBuf peek = inputRingBuffer.peek(receiveTakeBytes);
-                            if (peek != null && peek.readableBytes() >= 2 &&
-                                    peek.getByte(peek.readableBytes() - 2) == TTSConstant.TTS_END_BYTE &&
-                                    peek.getByte(peek.readableBytes() - 1) == TTSConstant.TTS_END_BYTE) {
-                                if (skipBytesInTheEndPacket != 0) skipEndData();
-                                peek.release();
-                                break;
-                            }
-                            if (peek != null) {
-                                peek.release();
-                            }
-                        }
-                        Thread.sleep(200);
-                        if (stop.get()) {
-                            log.debug("音频处理线程已停止");
-                            return;
-                        }
-                    }
-
-                    // 从输入缓冲区读取数据
-                    ByteBuf pcmData = takeDataAsByteBuf(inputRingBuffer, receiveTakeBytes);
-                    if (pcmData != null && pcmData.readableBytes() > 0) {
-                        // 先检查是否有结束标记，避免结束标记被编码产生噪音
-                        boolean hasEndFlag = false;
-                        if (pcmData.readableBytes() >= 2 &&
-                                pcmData.getByte(pcmData.readableBytes() - 2) == TTSConstant.TTS_END_BYTE &&
-                                pcmData.getByte(pcmData.readableBytes() - 1) == TTSConstant.TTS_END_BYTE) {
-                            hasEndFlag = true;
-                            // 去除结束标记，避免被编码为噪音
-                            pcmData.writerIndex(pcmData.writerIndex() - 2);
-                        }
-
-                        // 处理音频数据（重采样）
-                        ByteBuf processedData;
-                        if (reSample != null && reSample.equals("downsample24kTo8k")) {
-                            processedData = downsample24kTo8k(pcmData);
-                        } else {
-                            processedData = pcmData;
-                        }
-                        
-                        // 根据mediaType选择编码器
-                        ByteBuf g711Data = AudioCodecUtil.encode(processedData, mediaType);
-                        putData(outputRingBuffer, g711Data);
-                        g711Data.release();
-
-                        // 如果有结束标记，在编码后的数据之后添加未编码的结束标记
-                        if (hasEndFlag) {
-                            putData(outputRingBuffer, TTSConstant.TTS_END_FLAG.retainedDuplicate());
-                        }
-
-                        // 释放资源
-                        if (processedData != pcmData) {
-                            processedData.release();
-                        }
-                        pcmData.release();
-                    }
-                } catch (Exception e) {
-                    log.error("处理音频数据异常", e);
-                }
-            }
-        }).start();
+    public void putData(byte[] data, int bytesRead) {
+        if (data == null || data.length == 0 || bytesRead <= 0) {
+            return;
+        }
+        ByteBuf byteBuf = ByteBufAllocator.DEFAULT.buffer(bytesRead);
+        byteBuf.writeBytes(data, 0, bytesRead);
+        inputRingBuffer.write(byteBuf);
+        byteBuf.release();
     }
 
     /**
-     * 将PCM字节流从24kHz降采样到8kHz (ByteBuf版本，避免内存拷贝)
+     * 上游直接写入 ByteBuf 形式的原始 PCM 数据到 inputRingBuffer。
+     */
+    public void putData(ByteBuf data) {
+        if (data == null || data.readableBytes() == 0) {
+            return;
+        }
+        inputRingBuffer.write(data);
+    }
+
+    /**
+     * 单次推进一轮处理流程：检查 input、必要时裁尾、重采样、按协商编码组织输出，并写入 output。
+     * 该方法不阻塞、不自建线程，供共享处理调度器周期性调用。
+     */
+    public void processOnce() {
+        if (stop.get()) {
+            return;
+        }
+        try {
+            if (inputRingBuffer.getSize() == 0) {
+                return;
+            }
+
+            if (inputRingBuffer.getSize() < receiveTakeBytes) {
+                ByteBuf peek = inputRingBuffer.peek(receiveTakeBytes);
+                try {
+                    if (peek == null || peek.readableBytes() < 2) {
+                        return;
+                    }
+                    boolean hasEndFlag = peek.getByte(peek.readableBytes() - 2) == TTSConstant.TTS_END_BYTE
+                            && peek.getByte(peek.readableBytes() - 1) == TTSConstant.TTS_END_BYTE;
+                    if (!hasEndFlag) {
+                        return;
+                    }
+                    if (skipBytesInTheEndPacket != 0) {
+                        skipEndData();
+                    }
+                } finally {
+                    if (peek != null) {
+                        peek.release();
+                    }
+                }
+            }
+
+            ByteBuf pcmData = takeDataAsByteBuf(inputRingBuffer, receiveTakeBytes);
+            if (pcmData == null || pcmData.readableBytes() == 0) {
+                return;
+            }
+
+            boolean hasEndFlag = false;
+            if (pcmData.readableBytes() >= 2
+                    && pcmData.getByte(pcmData.readableBytes() - 2) == TTSConstant.TTS_END_BYTE
+                    && pcmData.getByte(pcmData.readableBytes() - 1) == TTSConstant.TTS_END_BYTE) {
+                hasEndFlag = true;
+                pcmData.writerIndex(pcmData.writerIndex() - 2);
+            }
+
+            ByteBuf processedData = pcmData;
+            if ("downsample24kTo8k".equals(reSample)) {
+                processedData = downsample24kTo8k(pcmData);
+            }
+
+            ByteBuf sendPayloadData = encodeRequired ? AudioCodecUtil.encode(processedData, mediaType) : processedData.retainedDuplicate();
+            putData(outputRingBuffer, sendPayloadData);
+            sendPayloadData.release();
+
+            if (hasEndFlag) {
+                putData(outputRingBuffer, TTSConstant.TTS_END_FLAG.retainedDuplicate());
+            }
+
+            if (processedData != pcmData) {
+                processedData.release();
+            }
+            pcmData.release();
+        } catch (Exception e) {
+            log.error("处理音频数据异常", e);
+        }
+    }
+
+    /**
+     * 将 PCM 字节流从 24kHz 降采样到 8kHz，沿用旧版算法语义。
      */
     public static ByteBuf downsample24kTo8k(ByteBuf input) {
         if (input == null || input.readableBytes() == 0) {
             return ByteBufAllocator.DEFAULT.buffer(0);
         }
 
-        int sampleSize = 2; // 每个采样点16-bit，即2字节
-        int ratio = 3;      // 24kHz -> 8kHz，每3个采样点降为1个
+        int sampleSize = 2;
+        int ratio = 3;
         int totalSamples = input.readableBytes() / sampleSize;
         int newSamples = totalSamples / ratio;
-
         ByteBuf output = ByteBufAllocator.DEFAULT.buffer(newSamples * sampleSize);
 
         for (int i = 0; i < newSamples; i++) {
             int idx1 = i * ratio * sampleSize;
             int idx2 = idx1 + sampleSize;
             int idx3 = idx2 + sampleSize;
-
-            // 防止越界（尾部可能不足3个采样点）
             if (idx3 + 1 >= input.readableBytes()) {
                 break;
             }
 
-            // 读取3个采样值（每个采样是2字节，小端）
             int s1 = (input.getByte(idx1 + 1) << 8) | (input.getByte(idx1) & 0xFF);
             int s2 = (input.getByte(idx2 + 1) << 8) | (input.getByte(idx2) & 0xFF);
             int s3 = (input.getByte(idx3 + 1) << 8) | (input.getByte(idx3) & 0xFF);
 
             int avg = (s1 + s2 + s3) / 3;
+            if (avg > Short.MAX_VALUE) {
+                avg = Short.MAX_VALUE;
+            }
+            if (avg < Short.MIN_VALUE) {
+                avg = Short.MIN_VALUE;
+            }
 
-            // 限幅防止溢出
-            if (avg > Short.MAX_VALUE) avg = Short.MAX_VALUE;
-            if (avg < Short.MIN_VALUE) avg = Short.MIN_VALUE;
-
-            // 写入输出（小端）
             output.writeByte((byte) (avg & 0xFF));
             output.writeByte((byte) ((avg >> 8) & 0xFF));
         }
@@ -199,188 +230,102 @@ public class NettyTtsRtpProcessor {
         return output;
     }
 
+    /**
+     * 检测到结束标记时，按旧版语义在 inputRingBuffer 中裁掉末尾多余 PCM，并重新补回 end flag。
+     */
     private void skipEndData() {
-        // 检测到结束标志，去掉末尾skipEndBytes字节的音频
         int totalSize = inputRingBuffer.getSize();
         if (totalSize > skipBytesInTheEndPacket + 2) {
-            // 读取所有数据（包含结束标志）
             ByteBuf allData = inputRingBuffer.read(totalSize);
-            // 去掉末尾的结束标志（2字节）和skipEndBytes字节音频
             int keepBytes = allData.readableBytes() - skipBytesInTheEndPacket - 2;
             if (keepBytes > 0) {
                 ByteBuf keepData = allData.readSlice(keepBytes);
                 inputRingBuffer.write(keepData.retainedDuplicate());
-                log.debug("检测到结束标志，已去除末尾{}字节，保留{}字节", skipBytesInTheEndPacket, keepBytes);
+                log.info("检测到结束标志，已去除末尾{}字节，保留{}字节", skipBytesInTheEndPacket, keepBytes);
             }
-            // 添加结束标志
             inputRingBuffer.write(TTSConstant.TTS_END_FLAG.retainedDuplicate());
             allData.release();
         } else {
-            // 数据不足skipEndBytes，只保留结束标志
             inputRingBuffer.clear();
             inputRingBuffer.write(TTSConstant.TTS_END_FLAG.retainedDuplicate());
-            log.debug("检测到结束标志，剩余数据不足{}字节，已清空", skipBytesInTheEndPacket);
+            log.info("检测到结束标志，剩余{}字节，剩余数据不足{}字节，已清空", totalSize, skipBytesInTheEndPacket);
         }
     }
 
     /**
-     * 向输入缓冲区写入数据
-     *
-     * @param data      输入数据
-     * @param bytesRead
+     * 从环形缓冲区读取一段 ByteBuf 数据，保持与旧版相同的读取上限语义。
      */
-    public void putData(byte[] data, int bytesRead) {
-        if (data == null || data.length == 0) {
-            return;
+    private ByteBuf takeDataAsByteBuf(NettyAudioRingBuffer buffer, int maxLength) {
+        if (buffer.getSize() == 0) {
+            return null;
         }
-
-        // 直接创建ByteBuf并写入数据，避免额外的内存拷贝
-        ByteBuf byteBuf = ByteBufAllocator.DEFAULT.buffer(bytesRead);
-        byteBuf.writeBytes(data, 0 , bytesRead);
-        inputRingBuffer.write(byteBuf);
-        byteBuf.release();
+        int actualLength = Math.min(maxLength, buffer.getSize());
+        ByteBuf byteBuf = buffer.read(actualLength);
+        if (byteBuf.readableBytes() == 0) {
+            byteBuf.release();
+            return null;
+        }
+        return byteBuf;
     }
 
     /**
-     * 向输入缓冲区写入ByteBuf数据
-     *
-     * @param data 输入ByteBuf数据
+     * 将一段 ByteBuf 数据写入指定环形缓冲区，用于复用旧版 output 写入路径。
      */
-    public void putData(ByteBuf data) {
+    private void putData(NettyAudioRingBuffer buffer, ByteBuf data) {
         if (data == null || data.readableBytes() == 0) {
             return;
         }
-
-        // 直接写入ByteBuf，零拷贝
-        inputRingBuffer.write(data);
+        buffer.write(data);
     }
 
     /**
-     * 开始RTP发送
+     * 注册共享处理调度任务和共享发送调度任务。
      */
     public void startRtpSender() {
-        new Thread(() -> {
-            boolean sendSilence = true;
-            // 创建静音数据ByteBuf，避免使用array()方法
-            ByteBuf silenceData = ByteBufAllocator.DEFAULT.buffer(160);
-            byte[] silenceBytes = new byte[160];
-            Arrays.fill(silenceBytes, TTSConstant.TTS_SILENCE_BYTE);
-            silenceData.writeBytes(silenceBytes);
-
-            while (true) {
-                try {
-                    // 控制每次分包是160字节 * n
-                    ByteBuf peek = outputRingBuffer.peek(EMConstant.VOIP_SAMPLES_PER_FRAME * 1000);
-                    if (stop.get()) {
-                        sender.close();
-                        silenceData.release(); // 释放silenceData
-                        return;
-                    }
-
-                    if (peek != null && peek.readableBytes() > 0) {
-                        sendSilence = false;
-
-                        int packageCount = peek.readableBytes() / EMConstant.VOIP_SAMPLES_PER_FRAME;
-                        int redundantData = peek.readableBytes() % EMConstant.VOIP_SAMPLES_PER_FRAME;
-
-                        // 解决Index -1 out of bounds for length 1问题
-                        if (peek.readableBytes() == 1 && peek.getByte(0) == TTSConstant.TTS_END_BYTE) {
-                            peek.release();
-                            // 直接使用ByteBuf常量
-                            putData(outputRingBuffer, TTSConstant.TTS_END_FLAG.retainedDuplicate());
-                            continue;
-                        }
-
-                        if (!(peek.getByte(peek.readableBytes() - 2) == TTSConstant.TTS_END_BYTE) &&
-                                !(peek.getByte(peek.readableBytes() - 1) == TTSConstant.TTS_END_BYTE) &&
-                                redundantData != 0) {
-                            if (packageCount > 1) {
-                                packageCount = packageCount - 1;
-                            } else {
-                                packageCount = 1;
-                            }
-                        } else if (packageCount == 0) {
-                            packageCount = 1;
-                        }
-
-                        peek.release(); // 释放peek的ByteBuf
-
-                        // 使用ByteBuf版本避免内存拷贝
-                        ByteBuf payload = takeDataAsByteBuf(outputRingBuffer, EMConstant.VOIP_SAMPLES_PER_FRAME * packageCount);
-                        if (payload != null && payload.readableBytes() > 0) {
-                            // 先检查是否有结束/中断标记，避免标记被发送出去产生噪音
-                            boolean hasEndFlag = false;
-                            boolean hasInterruptFlag = false;
-                            
-                            if (payload.readableBytes() >= 2) {
-                                byte endByte1 = payload.getByte(payload.readableBytes() - 2);
-                                byte endByte2 = payload.getByte(payload.readableBytes() - 1);
-
-                                if (endByte1 == TTSConstant.TTS_END_BYTE && endByte2 == TTSConstant.TTS_END_BYTE) {
-                                    hasEndFlag = true;
-                                    // 去除结束标记，避免作为音频数据发送
-                                    payload.writerIndex(payload.writerIndex() - 2);
-                                } else if (endByte1 == TTSConstant.TTS_INTERRUPT_BYTE && endByte2 == TTSConstant.TTS_INTERRUPT_BYTE) {
-                                    hasInterruptFlag = true;
-                                    // 去除中断标记，避免作为音频数据发送
-                                    payload.writerIndex(payload.writerIndex() - 2);
-                                }
-                            }
-
-                            // 只发送音频数据（不包含标记）
-                            if (payload.readableBytes() > 0) {
-                                sender.sendFrame(payload);
-                            }
-
-                            // 处理结束标记
-                            if (hasEndFlag) {
-                                sendSilence = true;
-                                SipUtils.executeTask(() -> callback.apply("completed"));
-                                log.info("tts播放完成");
-                            }
-                            
-                            // 处理中断标记
-                            if (hasInterruptFlag) {
-                                sendSilence = true;
-                                SipUtils.executeTask(() -> callback.apply("interrupt"));
-                                log.info("tts语音流已经被打断");
-                            }
-                            
-                            payload.release();
-                        }
-                    } else if (sendSilence) {
-                        sender.sendFrame(silenceData);
-                    }
-                } catch (Exception e) {
-                    log.error("RTP发送异常", e);
+        if (processScheduler == null) {
+            processScheduler = SpringUtils.getBean(TtsProcessScheduler.class);
+        }
+        if (rtpScheduler == null) {
+            rtpScheduler = SpringUtils.getBean(TtsRtpScheduler.class);
+        }
+        if (processTaskId == null) {
+            processTaskId = processScheduler.register(this);
+        }
+        if (schedulerTaskId == null) {
+            schedulerTaskId = rtpScheduler.register(this, result -> {
+                if (callback != null) {
+                    callback.apply(result);
                 }
-            }
-        }).start();
+            });
+        }
     }
 
     /**
-     * 停止RTP发送
+     * 停止处理器并取消处理/发送调度任务。
      */
     public void stopRtpSender() {
         stop.set(true);
+        if (processScheduler != null && processTaskId != null) {
+            processScheduler.cancel(processTaskId);
+            processTaskId = null;
+        }
+        if (rtpScheduler != null && schedulerTaskId != null) {
+            rtpScheduler.cancel(schedulerTaskId);
+            schedulerTaskId = null;
+        }
         if (sender != null) {
             sender.close();
         }
-        // 释放缓冲区资源
         releaseResources();
     }
 
     /**
-     * 中断TTS播放
+     * 沿用旧版 interrupt 语义：同时清 input/output，并把 interrupt 标记写入 output。
      */
     public void interrupt() {
         try {
             inputRingBuffer.clear();
             outputRingBuffer.clear();
-            if (sender != null) {
-                sender.interrupt();
-            }
-            // 直接使用ByteBuf常量
             putData(outputRingBuffer, TTSConstant.TTS_INTERRUPT_FLAG.retainedDuplicate());
             log.debug("已中断TTS播放");
         } catch (Exception e) {
@@ -389,79 +334,18 @@ public class NettyTtsRtpProcessor {
     }
 
     /**
-     * 辅助方法：从NettyCircularAudioBuffer读取数据，返回ByteBuf
-     */
-    private ByteBuf takeDataAsByteBuf(NettyAudioRingBuffer buffer, int maxLength) {
-        if (buffer.getSize() == 0) {
-            return null;
-        }
-
-        int actualLength = Math.min(maxLength, buffer.getSize());
-        ByteBuf byteBuf = buffer.read(actualLength);
-
-        if (byteBuf.readableBytes() == 0) {
-            byteBuf.release();
-            return null;
-        }
-
-        return byteBuf;
-    }
-
-    /**
-     * 辅助方法：向NettyCircularAudioBuffer写入ByteBuf数据
-     */
-    private void putData(NettyAudioRingBuffer buffer, ByteBuf data) {
-        if (data == null || data.readableBytes() == 0) {
-            return;
-        }
-
-        // 直接写入ByteBuf，零拷贝
-        buffer.write(data);
-    }
-
-    /**
-     * 释放所有资源
+     * 释放 input/output ring buffer 资源。
      */
     public void releaseResources() {
         try {
-            if (inputRingBuffer != null && !inputRingBuffer.isClosed()) {
+            if (!inputRingBuffer.isClosed()) {
                 inputRingBuffer.release();
-                log.debug("已释放输入缓冲区资源");
             }
-            if (outputRingBuffer != null && !outputRingBuffer.isClosed()) {
+            if (!outputRingBuffer.isClosed()) {
                 outputRingBuffer.release();
-                log.debug("已释放输出缓冲区资源");
             }
         } catch (Exception e) {
             log.warn("释放缓冲区资源时出现异常", e);
         }
     }
-
-    /**
-     * 获取缓冲区状态信息
-     */
-    public String getBufferStatus() {
-        return String.format("TTS缓冲区状态 - 输入：%s, 输出：%s",
-                inputRingBuffer.getStatusInfo(),
-                outputRingBuffer.getStatusInfo());
-    }
-
-    /**
-     * 检查缓冲区健康状态
-     */
-    private void checkBufferHealth() {
-        double inputUsage = inputRingBuffer.getUsageRatio();
-        double outputUsage = outputRingBuffer.getUsageRatio();
-
-        if (inputUsage > 0.9) {
-            log.warn("输入缓冲区使用率过高：{:.1f}%，TTS数据积压！", inputUsage * 100);
-        }
-        if (outputUsage > 0.9) {
-            log.warn("输出缓冲区使用率过高：{:.1f}%，RTP发送速度跟不上！", outputUsage * 100);
-        }
-
-        if (inputUsage > 0.8 || outputUsage > 0.8) {
-            log.debug("缓冲区状态：{}", getBufferStatus());
-        }
-    }
-} 
+}
