--- src/main/java/com/cfsl/easymrcp/tts/NettyTtsRtpProcessor3.java
@@ -1,111 +1,0 @@
-package com.cfsl.easymrcp.tts;
-
-import com.cfsl.easymrcp.common.EMConstant;
-import com.cfsl.easymrcp.mrcp.TtsCallback;
-import com.cfsl.easymrcp.rtp.NettyAudioRingBuffer;
-import com.cfsl.easymrcp.rtp.NettyRtpSenderV3;
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.ByteBufAllocator;
-import io.netty.channel.Channel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.extern.slf4j.Slf4j;
-
-/**
- * 重构后的 TTS 音频处理器（V3）。
- * 保留该文件用于对照旧版简化实现，但当前主链路已经切到 V4，
- * 因此这里不再接入共享发送调度器。
- */
-@Slf4j
-public class NettyTtsRtpProcessor3 {
-
-    @Getter
-    private NettyRtpSenderV3 sender;
-    @Setter
-    private TtsCallback callback;
-
-    private static final int TTS_OUTPUT_BUFFER_SECONDS = 30;
-    private static final int SAMPLE_RATE = EMConstant.VOIP_SAMPLE_RATE;
-
-    @Getter
-    private final NettyAudioRingBuffer outputRingBuffer;
-
-    @Setter
-    @Getter
-    private int skipBytesInTheEndPacket = 0;
-
-    public NettyTtsRtpProcessor3(String remoteIp, int remotePort, int mediaType) {
-        ByteBufAllocator allocator = ByteBufAllocator.DEFAULT;
-        this.outputRingBuffer = new NettyAudioRingBuffer(allocator, SAMPLE_RATE, TTS_OUTPUT_BUFFER_SECONDS, true);
-
-        try {
-            this.sender = new NettyRtpSenderV3(remoteIp, remotePort);
-            this.sender.setPayloadType(mediaType);
-        } catch (Exception e) {
-            log.error("初始化 NettyTtsRtpProcessorV3 失败", e);
-            throw new RuntimeException("初始化 RTP 发送器失败", e);
-        }
-    }
-
-    public void setRtpChannel(Channel channel) {
-        sender.setRtpChannel(channel);
-    }
-
-    public void putData(byte[] data, int bytesRead) {
-        if (data == null || data.length == 0 || bytesRead <= 0) {
-            return;
-        }
-        ByteBuf byteBuf = ByteBufAllocator.DEFAULT.buffer(bytesRead);
-        byteBuf.writeBytes(data, 0, bytesRead);
-        outputRingBuffer.write(byteBuf);
-        byteBuf.release();
-    }
-
-    public void putData(ByteBuf data) {
-        if (data == null || data.readableBytes() == 0) {
-            return;
-        }
-        outputRingBuffer.write(data);
-    }
-
-    public void writeEndFlag() {
-        outputRingBuffer.write(TTSConstant.TTS_END_FLAG.retainedDuplicate());
-    }
-
-    public void startRtpSender() {
-        log.warn("NettyTtsRtpProcessor3 仅保留代码对照用途，当前版本不再接入发送调度器");
-    }
-
-    public void stopRtpSender() {
-        if (sender != null) {
-            sender.close();
-        }
-        releaseResources();
-    }
-
-    public void interrupt() {
-        try {
-            outputRingBuffer.clear();
-            if (sender != null) {
-                outputRingBuffer.write(TTSConstant.TTS_INTERRUPT_FLAG.retainedDuplicate());
-            }
-            log.debug("已中断 TTS 播放");
-        } catch (Exception e) {
-            log.error("中断 TTS 播放时出现异常", e);
-        }
-    }
-
-    public void releaseResources() {
-        try {
-            if (outputRingBuffer != null && !outputRingBuffer.isClosed()) {
-                outputRingBuffer.release();
-            }
-        } catch (Exception e) {
-            log.warn("释放缓冲区资源时出现异常", e);
-        }
-    }
-
-    public String getBufferStatus() {
-        return String.format("TTS V3 - 缓冲区: %s", outputRingBuffer.getStatusInfo());
-    }
-}
